requirements.txt
emoji==2.2.0
fastapi==0.102.0
pandas==1.3.4
python-dotenv==1.0.0
transformers==4.21.2
tweepy==4.12.1
